Add TransactionPurchaseRepository(Test) (My Ky, Maren)

+++ src/main/java/de/hhu/sharing/data/TransactionPurchaseRepository.java
@@ -2,12 +2,11 @@ package de.hhu.sharing.data;
 
 import de.hhu.sharing.model.User;
 import de.hhu.sharing.propay.TransactionPurchase;
-import de.hhu.sharing.propay.TransactionRental;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface TransactionPurchaseRepository extends CrudRepository<TransactionPurchase, Integer> {
+public interface TransactionPurchaseRepository extends CrudRepository<TransactionPurchase, Long> {
     List<TransactionPurchase> findAll();
     List<TransactionPurchase> findAllBySender(User user);
     List<TransactionPurchase> findAllByReceiver(User user);
